Fix modification of bimps in look while iterating over bimps
While looking ahead, we can find "windfalls" that imply new binary clauses.
There was a code path that tried to add these windfalls to the bimp vector
while iteration bimp was ongoing, which could cause a segfault if bimp gets
resized by the additions. Fixing this by adding the windfalls higher up in
the call stack.